move subgraph shared handlers to plugin folder
