+++ apps/ensindexer/src/plugins/subgraph/plugins/basenames/event-handlers.ts
@@ -4,7 +4,7 @@ import { PluginName } from "@ensnode/ensnode-sdk";
 import attach_Registrar from "./handlers/Registrar";
 import attach_Registry from "./handlers/Registry";
 
-import attach_SharedMultichainResolverHandlers from "@/handlers/multi-chain/Resolver";
+import attach_SharedMultichainResolverHandlers from "@/plugins/subgraph/shared-handlers/multi-chain/Resolver";
 
 // conditionally attach event handlers when Ponder executes this file
 if (config.plugins.includes(PluginName.Basenames)) {
