+++ apps/ensindexer/src/plugins/subgraph/plugins/lineanames/event-handlers.ts
@@ -1,7 +1,7 @@
 import config from "@/config";
 import { PluginName } from "@ensnode/ensnode-sdk";
 
-import attach_SharedMultichainResolverHandlers from "@/handlers/multi-chain/Resolver";
+import attach_SharedMultichainResolverHandlers from "@/plugins/subgraph/shared-handlers/multi-chain/Resolver";
 
 import attach_NameWrapper from "./handlers/NameWrapper";
 import attach_Registrar from "./handlers/Registrar";
