+++ apps/ensindexer/src/plugins/subgraph/plugins/lineanames/handlers/NameWrapper.ts
@@ -1,8 +1,8 @@
 import { ponder } from "ponder:registry";
 
 import config from "@/config";
-import { makeNameWrapperHandlers } from "@/handlers/NameWrapper";
 import { namespaceContract } from "@/lib/plugin-helpers";
+import { makeNameWrapperHandlers } from "@/plugins/subgraph/shared-handlers/NameWrapper";
 import { PluginName } from "@ensnode/ensnode-sdk";
 import { getRegistrarManagedName } from "../lib/registrar-helpers";
