+++ apps/ensindexer/src/plugins/subgraph/shared-handlers/Registry.ts
